requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
Django==2.1.5
django-cors-headers==2.4.0
django-debug-permissions==1.0.0
django-extensions==2.1.5
django-mailgun==0.9.1
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
et-xmlfile==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
livereload==2.6.1
lxml==4.3.3
Markdown==3.0.1
MarkupSafe==1.1.0
mkdocs==1.0.4
mkdocs-material==4.3.1
openpyxl==2.6.1
Pillow==6.0.0
pip-autoremove==0.9.1
psycopg2==2.8.2
pydotplus==2.0.2
Pygments==2.4.1
pymdown-extensions==6.0
pyparsing==2.3.1
PyQRCode==1.2.1
pytz==2018.9
PyYAML==3.13
reportlab==3.5.21
requests==2.21.0
six==1.12.0
svglib==0.9.0
tinycss2==1.0.2
tornado==6.0.2
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.18
webencodings==0.5.1
XlsxWriter==1.1.5
